requirements.txt
alembic==1.8.1
click==8.1.3
Flask==2.2.2
flask-marshmallow==0.14.0
Flask-Migrate==3.1.0
Flask-SQLAlchemy==2.5.1
greenlet==1.1.3
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.1
MarkupSafe==2.1.1
marshmallow==3.17.1
packaging==21.3
PyMySQL==1.0.2
pyparsing==3.0.9
six==1.16.0
SQLAlchemy==1.4.40
Werkzeug==2.2.2
python-dotenv==0.20.0
apifairy==1.0.0